+++ public/index.html
@@ -23,33 +23,8 @@
 </header>
 
 <main class="main">
-  <nav class="main-navigation">
-  </nav>
 
-  <section class="films">
-    <section class="films-list">
-      <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
 
-      <div class="films-list__container films-list__container--main">
-      </div>
-
-      <button class="films-list__show-more">Show more</button>
-    </section>
-
-    <section class="films-list--extra">
-      <h2 class="films-list__title">Top rated</h2>
-
-      <div class="films-list__container films-list__container--top-rated">
-      </div>
-    </section>
-
-    <section class="films-list--extra">
-      <h2 class="films-list__title">Most commented</h2>
-
-      <div class="films-list__container films-list__container--most-commented">
-      </div>
-    </section>
-  </section>
 </main>
