+++ src/get-movies.js
@@ -1,4 +1,4 @@
-import generateMovie from './generate-movie.js';
+import generateMovie from './helpers/generate-movie.js';
 
 export default (count) => {
   return new Array(count).fill(null).map(() => generateMovie());
