--- src/mock-data.js
@@ -1,49 +1,0 @@
-import {PATH_TO_IMG} from './constants.js';
-
-export default {
-  title: [
-    `Аквамен`,
-    `Апгрейд`,
-    `Хищник`,
-    `Властелин колец`,
-    `Веном`,
-    `Мстители`,
-    `Интерстеллар`,
-    `Хоббит`,
-    `Звездные войны`,
-    `Начало`,
-    `Щерлок`,
-    `Зеленая миля`,
-    `Великий Гетсби`,
-    `Планета обезьян`,
-    `Тор`,
-  ],
-  pictures: [
-    `${PATH_TO_IMG}accused.jpg`,
-    `${PATH_TO_IMG}blackmail.jpg`,
-    `${PATH_TO_IMG}blue-blazes.jpg`,
-    `${PATH_TO_IMG}fuga-da-new-york.jpg`,
-    `${PATH_TO_IMG}moonrise.jpg`,
-    `${PATH_TO_IMG}three-friends.jpg`,
-  ],
-  sentences: [
-    `Lorem ipsum dolor sit amet, consectetur adipiscing elit.`,
-    `Cras aliquet varius magna, non porta ligula feugiat eget.`,
-    `Fusce tristique felis at fermentum pharetra.`,
-    `Aliquam id orci ut lectus varius viverra.`,
-    `Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante.`,
-    `Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum.`,
-    `Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui.`,
-    `Sed sed nisi sed augue convallis suscipit in sed felis.`,
-    `Aliquam erat volutpat.`,
-    `Nunc fermentum tortor ac porta dapibus. In rutrum ac purus sit amet tempus.`,
-  ],
-  genres: [
-    `Comedy`,
-    `Drama`,
-    `Horror`,
-    `Thriller`,
-    `Action`,
-    `Fantasy`
-  ],
-};
